DLS-6253: Tidied up and split out ReportingEntityData value types
